+++ CHANGELOG.md
@@ -4,6 +4,23 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 
 ## [{version}] - {release_date}
 ### Added
+- Generates a multiple line billing log file csv based on accepted papers that integrates EM GUID and
+  information from various cards
+### Changed
+### Deprecated
+### Removed
+- Removed University of Ottawa as an institutional account participant
+- Invitations can no longer be assigned via invitation codes.
+### Fixed
+- Fixed figures in PDFs spanning multiple pages
+### Security
+
+## [1.4.11] - 2016-05-19
+### Fixed
+- Fixed errors during PDF generation
+
+## [1.4.10] - 2016-05-18
+### Added
 - Diffing of Reviewer Candidates
 - Authors will see the feedback form whenever they submit a paper
 - Add manifest to Apex export
@@ -12,6 +29,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 - Update RTC and FTC text.
 - Billing task institutions and countries synced with EM
 
+## [1.4.9] - 2016-05-13
 ### Deprecated
 ### Removed
 ### Fixed